# animal-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the veterinary API",
        "title": "Veterinary API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/cats/{catid}/visits": {
            "get": {
                "description": "Get all visits for a specific cat",
                "tags": [
                    "visits"
                ],
                "summary": "Get all visits",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "catid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Visit"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a visit for a specific cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Create a visit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "catid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Visit info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Visit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Visit"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cats/{catid}/visits/{id}": {
            "get": {
                "description": "Get a visit by ID and for a specific cat",
                "tags": [
                    "visits"
                ],
                "summary": "Get a visit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "catid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Visit"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a visit for a specific cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Update a visit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "catid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Visit info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Visit"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Cat": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "Visit": {
            "type": "object",
            "properties": {
                "cat": {
                    "$ref": "#/definitions/Cat"
                },
                "date": {
                    "type": "string"
                }
            }
        }
    }
}